[Parent][MBL-13832] Webview helpers
